requirements.txt
GitPython >= 3.0.5
awscli >= 1.16.302
beautifulsoup4 >= 4.8.0
boto3 >= 1.10.38
html5lib >= 1.0.1
mypy >= 0.730
networkx >= 2.4
pdfminer.six == 20181108
pytextrank >= 2.0.0
ray >= 0.6.5
rdflib >= 4.2.2
rdflib-jsonld >= 0.4.0
requests-html >= 0.10.0
spacy >= 2.2.3
sxsdiff >= 0.3.0
tqdm >= 4.37.0
